Separate out UI code
The entire code for the rendering of the application UI is
all inside the Duke start method.

Let's move this code to the Ui class to keep the chatbot
logic and UI separate.
