+++ src/main/java/duke/commands/ListCommand.java
@@ -8,6 +8,6 @@ public class ListCommand extends Command {
     @Override
     public void run(Parser parser, Duke duke) throws DukeOptionParsingException {
         parser.assertEnd();
-        duke.print("Here's what you've done today...\n" + duke.getTasks().toDisplayString());
+        duke.getUi().print("Here's what you've done today...\n" + duke.getTasks().toDisplayString());
     }
 }
