+++ src/main/java/duke/ui/Ui.java
@@ -1,37 +1,118 @@
 package duke.ui;
 
-import java.util.Scanner;
+import duke.Duke;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.Region;
+import javafx.scene.layout.VBox;
+import javafx.stage.Stage;
+
+import java.util.Objects;
 
 public class Ui {
 
-    /** Scanner for user input. */
-    private final Scanner input = new Scanner(System.in);
+    private ScrollPane scrollPane;
+    private VBox dialogContainer;
+    private TextField userInput;
+    private Button sendButton;
+    private Scene scene;
+
+    private Image userImage = new Image(Objects.requireNonNull(this.getClass().getResourceAsStream("/images/user.jpg")));
+    private Image dukeImage = new Image(Objects.requireNonNull(this.getClass().getResourceAsStream("/images/louie.jpg")));
+
+    private Duke duke;
+
+    public Ui(Stage stage, Duke duke) {
+        //Step 1. Setting up required components
+        this.duke = duke;
+
+        //The container for the content of the chat to scroll.
+        scrollPane = new ScrollPane();
+        dialogContainer = new VBox();
+        scrollPane.setContent(dialogContainer);
+
+        userInput = new TextField();
+        sendButton = new Button("Send");
+
+        AnchorPane mainLayout = new AnchorPane();
+        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
+
+        scene = new Scene(mainLayout);
+
+        stage.setScene(scene);
+        stage.show();
+
+        //More code to be added here later
+        //Step 2. Formatting the window to look as expected
+        stage.setTitle("Duke");
+        stage.setResizable(false);
+        stage.setMinHeight(600.0);
+        stage.setMinWidth(400.0);
+
+        mainLayout.setPrefSize(400.0, 600.0);
+
+        scrollPane.setPrefSize(385, 535);
+        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
+        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
+
+        scrollPane.setVvalue(1.0);
+        scrollPane.setFitToWidth(true);
+
+        //You will need to import `javafx.scene.layout.Region` for this.
+        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
+
+        userInput.setPrefWidth(325.0);
+
+        sendButton.setPrefWidth(55.0);
+
+        AnchorPane.setTopAnchor(scrollPane, 1.0);
+
+        AnchorPane.setBottomAnchor(sendButton, 1.0);
+        AnchorPane.setRightAnchor(sendButton, 1.0);
+
+        AnchorPane.setLeftAnchor(userInput , 1.0);
+        AnchorPane.setBottomAnchor(userInput, 1.0);
+
+        //Part 3. Add functionality to handle user input.
+        sendButton.setOnMouseClicked((event) -> {
+            handleUserInput();
+        });
+
+        userInput.setOnAction((event) -> {
+            handleUserInput();
+        });
+
+        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
+    }
 
     /**
-     * Prints a message to the terminal, decorated with the Louie icon.
-     * Printing an empty string just outputs the icon
-     *
-     * @param message The message to print.
+     * Prints message to the chatbot GUI.
+     * @param msg the message to be printed.
      */
-    public void print(String message) {
-        boolean isFirst = true;
-        StringBuilder sb = new StringBuilder();
-        sb.append("\n");
-        for (String s : message.split("\n")) {
-            sb.append("    ");
-            if (isFirst) {
-                sb.append("(>^.^<)");
-                isFirst = false;
-            } else {
-                sb.append("       ");
-            }
-            sb.append(" ").append(s).append("\n");
-        }
-        sb.append("\n");
-        System.out.print(sb);
+    public void print(String msg) {
+        dialogContainer.getChildren()
+                .add(DialogBox.createDukeDialog
+                        (new Label(msg), new ImageView(dukeImage)));
+
     }
 
-    public String readInput() {
-        return input.nextLine();
+    /**
+     * Iteration 2:
+     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
+     * the dialog container. Clears the user input after processing.
+     */
+    private void handleUserInput() {
+        String readString = userInput.getText();
+        Label userLabel = new Label(readString);
+        dialogContainer.getChildren().add(DialogBox.createUserDialog(userLabel, new ImageView(userImage)));
+        userInput.clear();
+        this.duke.processInput(readString);
     }
+
 }
